move filterchecking into separate utility
