+++ src/components/utils/index.tsx
@@ -1,5 +1,6 @@
 import useRedirectTimer from "./useRedirectTimer";
 import { getTodaysDate, getTomorrowsDate } from "./dateUtils";
 import useScrollToTop from "./useScrollToTop";
+import { checkForFilterChanges, setInitialFilterValues } from "./filterVenuesUtils";
 
-export { useRedirectTimer, getTodaysDate, getTomorrowsDate, useScrollToTop };
+export { useRedirectTimer, getTodaysDate, getTomorrowsDate, useScrollToTop, checkForFilterChanges, setInitialFilterValues };
